refactor(ports): relocate shared ports from kernel/ports/ to hiresense/ports/
Shared Protocol interfaces (EventBus, LLMPort, VectorStorePort, LaTeXCompilerPort)
now live at top-level hiresense/ports/ per the kernel-and-shared-types standard.
kernel/ports/ converted to backward-compatible re-export shims.
