+++ backend/src/hiresense/kernel/ports/event_bus.py
@@ -1,16 +1,5 @@
-from __future__ import annotations
+"""Backward-compatible re-export. Import from hiresense.ports instead."""
 
-from collections.abc import Awaitable, Callable
-from typing import Protocol
+from hiresense.ports.event_bus import EventBus
 
-from hiresense.kernel.events import DomainEvent
-
-
-class EventBus(Protocol):
-    async def publish(self, event: DomainEvent) -> None: ...
-
-    def subscribe(
-        self,
-        event_type: str,
-        handler: Callable[[DomainEvent], Awaitable[None]],
-    ) -> None: ...
+__all__ = ["EventBus"]
