+++ backend/src/hiresense/kernel/ports/latex_compiler.py
@@ -1,16 +1,5 @@
-from __future__ import annotations
+"""Backward-compatible re-export. Import from hiresense.ports instead."""
 
-from dataclasses import dataclass
-from typing import Protocol
+from hiresense.ports.latex_compiler import CompilationError, LaTeXCompilerPort
 
-
-@dataclass
-class CompilationError:
-    line: int
-    message: str
-
-
-class LaTeXCompilerPort(Protocol):
-    async def compile(self, tex_content: str) -> bytes: ...
-
-    async def validate(self, tex_content: str) -> list[CompilationError]: ...
+__all__ = ["CompilationError", "LaTeXCompilerPort"]
