+++ backend/src/hiresense/ports/__init__.py
@@ -1,0 +1,13 @@
+from hiresense.ports.event_bus import EventBus
+from hiresense.ports.latex_compiler import CompilationError, LaTeXCompilerPort
+from hiresense.ports.llm import LLMPort
+from hiresense.ports.vector_store import ScoredResult, VectorStorePort
+
+__all__ = [
+    "CompilationError",
+    "EventBus",
+    "LaTeXCompilerPort",
+    "LLMPort",
+    "ScoredResult",
+    "VectorStorePort",
+]
