+++ backend/src/hiresense/ports/event_bus.py
@@ -1,0 +1,16 @@
+from __future__ import annotations
+
+from collections.abc import Awaitable, Callable
+from typing import Protocol
+
+from hiresense.kernel.events import DomainEvent
+
+
+class EventBus(Protocol):
+    async def publish(self, event: DomainEvent) -> None: ...
+
+    def subscribe(
+        self,
+        event_type: str,
+        handler: Callable[[DomainEvent], Awaitable[None]],
+    ) -> None: ...
